allow for maximum file size for files:pull
a new constant MAX_FILE_PULL_SIZE was added to the config file. This is
appended to a --max-size parameter for the rsync command when pulling
down a copy of the files.

this was added to prevent huge files from being pulled down to a local
development copy of a site. If the constant is not set it will be ignored
and files of any size will be pulled down.